pyproject.toml
[project]
name = "classawarepruning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "torch>=2.7.1",
    "torch-pruning>=1.5.3",
    "torchvision>=0.22.1",
]
